Repository restructuring with replacing locators and pageURLs objects to pageElements/loginPage
