+++ tests/basicAuth.spec.js
@@ -1,6 +1,6 @@
 // @ts-check
 const { test } = require('@playwright/test');
-const { pageData, locators } = require('../fixtures/commonData');
+const { locators, pageURLs,isLogin } = require('../pageElements/loginPage');
 
 test.use({ extraHTTPHeaders: {
     Authorization: 'Basic dXNlcjpwYXNz'
@@ -10,5 +10,5 @@ test('basic auth with auth header', async ({ page }) => {
     await page.goto('/')
     await page.getByText('HTTP/NTLM Auth', { exact: true }).click();
 
-    await pageData.isLogin(page, locators.login.successUrl, locators.login.successText);
+    await isLogin(page, pageURLs.successUrl, locators.successText);
 })
